Move Block specific tests to new test suite
